refactor(_apiref): fold the _griffe package into _docstring_sections
The directory name suggested a vendored copy of griffe internals, but
since the shim was removed (486a91e-era) it held only the project-owned
dataclass docstring sections. The DC prefix stays: it marks sections
for documenting dataclasses (init parameters, parameter attributes),
not a copied-from-griffe distinction.
