--- great_docs/_apiref/_griffe/__init__.py
@@ -1,18 +1,0 @@
-"""
-great-docs extensions to griffe's docstring model
-
-Only genuine additions live here; for everything griffe already provides,
-import `griffe` directly.
-"""
-
-from .docstrings import (
-    DCDocstringSectionInitParameters,
-    DCDocstringSectionParameterAttributes,
-)
-from .enumerations import DCDocstringSectionKind
-
-__all__ = (
-    "DCDocstringSectionKind",
-    "DCDocstringSectionParameterAttributes",
-    "DCDocstringSectionInitParameters",
-)
